Use a more functional style to create gates
The result of which is we pass around "naked" event emitters, instead of a class wrapping event emitters. I think this makes the code nicer and more flexible. It does, however, introduce a bit of ceremony around typing. For example, we now have things like:

import type { Type as Wire } from './Wire';

That doesn't seem that bad, though.